praxis_ai/tools/web_search.py
# tools/web_search.py

import requests
from bs4 import BeautifulSoup
from ..utils.logging import logger

def web_search(query: str, num_results: int = 5) -> list:
    """
    Perform a web search and return a list of search results.
    This is a placeholder implementation and should be replaced with a proper search API.
    """
    try:
        # This is a mock implementation
        results = [
            {"title": f"Result {i} for {query}", "url": f"https://example.com/result{i}", "snippet": f"This is a snippet for result {i}"}
            for i in range(1, num_results + 1)
        ]
        return results
    except Exception as e:
        logger.error(f"Error performing web search: {e}")
        return []

def extract_content(url: str) -> str:
    """
    Extract the main content from a web page.
    This is a basic implementation and may need to be improved for production use.
    """
    try:
        response = requests.get(url)
        soup = BeautifulSoup(response.content, 'html.parser')
        
        # Remove script and style elements
        for script in soup(["script", "style"]):
            script.decompose()
        
        # Get text
        text = soup.get_text()
        
        # Break into lines and remove leading and trailing space on each
        lines = (line.strip() for line in text.splitlines())
        
        # Break multi-headlines into a line each
        chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
        
        # Drop blank lines
        text = '\n'.join(chunk for chunk in chunks if chunk)
        
        return text
    except Exception as e:
        logger.error(f"Error extracting content from {url}: {e}")
        return ""